clinical_trials/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "This is a data pipelining server for clinical trials data"
authors = ["Rathijit Paul <rathijit.paul@curieo.org>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.29"
asyncio = "^3.4.3"
requests = "^2.31.0"
bs4 = "^0.0.2"
python-dotenv = "^1.0.1"
tqdm = "^4.66.2"
llama-index = "^0.10.27"
llama-index-embeddings-text-embeddings-inference = "^0.1.2"
llama-index-vector-stores-qdrant = "^0.1.4"
llama-index-llms-together = "^0.1.3"
llama-index-readers-database = "^0.1.3"
psycopg2-binary = "^2.9.9"

pydantic-settings = "^2.2.1"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
main = "app.main:main"
